# [8_11] boost performance of api converting number to hexadecimal string