--- a/lolly/data/numeral.cpp
+++ b/lolly/data/numeral.cpp
@@ -156,20 +156,26 @@ to_padded_hex (uint8_t i) {
 
 template <typename T>
 std::enable_if_t<std::conjunction_v<std::is_integral<T>, std::is_unsigned<T>>,
-                 string>
-to_Hex_positive (T i) {
-  if (i < 16) return hex_string[i & 15];
-  return to_Hex_positive (i >> 4) << hex_string[i & 15];
+                 void>
+to_Hex_positive (T i, string& s) {
+  if (i >= 16) {
+    to_Hex_positive (i >> 4, s);
+  }
+  s << hex_string[i & 15];
 }
 
 string
 to_Hex (int32_t i) {
   if (i == INT32_MIN) return "-80000000";
   if (i < 0) {
-    return string ("-") << to_Hex_positive ((uint32_t) (-i));
+    string result ("-");
+    to_Hex_positive ((uint32_t) (-i), result);
+    return result;
   }
   else {
-    return to_Hex_positive ((uint32_t) i);
+    string result;
+    to_Hex_positive ((uint32_t) (i), result);
+    return result;
   };
 }
 
@@ -182,10 +188,14 @@ string
 to_Hex (pointer ptr) {
   intptr_t i= (intptr_t) ptr;
   if (i < 0) {
-    return string ("-") << to_Hex_positive ((uintptr_t) (-i));
+    string result ("-");
+    to_Hex_positive ((uintptr_t) (-i), result);
+    return result;
   }
   else {
-    return to_Hex_positive ((uintptr_t) i);
+    string result;
+    to_Hex_positive ((uintptr_t) (i), result);
+    return result;
   };
 }
 
@@ -207,10 +217,21 @@ from_hex (string s) {
   return res;
 }
 
+template <typename T>
+std::enable_if_t<std::conjunction_v<std::is_integral<T>, std::is_unsigned<T>>,
+                 void>
+as_hexadecimal_sub (T i, unsigned int cur, string& s) {
+  if (cur > 0) {
+    as_hexadecimal_sub (i >> 4, cur - 1, s);
+  }
+  s[cur]= hex_string[i & 15];
+}
+
 string
 as_hexadecimal (int i, int len) {
-  if (len == 1) return hex_string[i & 15];
-  else return as_hexadecimal (i >> 4, len - 1) << hex_string[i & 15];
+  string result (len);
+  as_hexadecimal_sub ((unsigned) i, len - 1, result);
+  return result;
 }
 
 } // namespace data